Drop websockets dependency
Previously, rpcx.websockets was the only consumer of websockets.

While the contents of the module were useful, there were limitations
with the implementation and the dependency on websockets became complex
with changes to the library to not provide a public async interface that
could be used for consistent typing across the legacy implementation and
newer implmentations and that were stable across versions.

As such, it makes sense to drop the module but to provide an example of
how users may take advantage of this library for use with `RPCClient`
and `RPCServer`.

Signed-off-by: Vincent Fazio <vfazio@xes-inc.com>