client: abort remaining requests after the server stream ends
If the client's server receive stream is closed from the other end while
a request is waiting for a response, the request will hang indefinitely since
the memory object stream pairs in `_RequestTask` are never closed out.

Following the end of `receive_loop`, add handling to close the producer streams
of any remaining request tasks so that unresponsive requests or request streams
raise an EndOfStream exception instead of hanging indefinitely.